Theme and Domain data
this is all the Theme/Domain data on the cards (except for the icons), plus the long descriptions from the book.

i've left the shared entries (i.e. 89-98, 99, 100 on Domains) intact.

they probably won't be useful as-is as rolled oracles, because i've opted to keep "Chance" as the top of the roll range for the sake of convention, but i can't really think of a reason that'd be necessary in play.